address comment (#168)
